Adding grafana test function
This sets `/test` as the te route to use for testing a DataSource
connection status. Previously `/` was being used, but that route
expects some additional params causing a failure.

I've additionally added a check to buildQueryParameters. If any of the
targets are plain strings the target is returned; This resolves a
failing test.