Refactoring to allow arbitrary targets
Previously we recomputed the target name for each tsds result key and
attempted a datapoint lookup to find any assoicated datapoints. If a
target name and assoicated datapoints were found we added the data to
our result.

We now return the original tsds datapoint name along with each new
target name in `findtarget_names`. Using the original names, we do a
lookup for each and return their associated datapoints.