Adding support for adhoc filters
This change adds support for adhoc filters to GenericDatasource.

When defining an adhoc template variable, the name used will be used
as the tsds measurement type; This will determine which filters are
available to the user.

When building a panel, be aware that the selected filters will be
appened to the where clause as defined in the query builder.

Also note that template variables or other adhoc filters will be used
to filter the values of subsquent adhoc filters. For example, if you
first filter on intf=et-0/0/0, only nodes with an et-0/0/0 interface
will be available under the node filter.