Adding ability to use sub function in raw query
Using functions like `sum(aggregate(values.output, 300, average))` in
a raw query, would cause the python middleware to crash while building
graph labels. We now check that the top level function is aggregate;
If it's not, we generate the standard graph label based on specified
metrics.

This change also stops the middleware from crashing when no target
names are specified.

Also re-adding a default bucket size when not specified by the user.