Changing buildQueryParameters to return a Promise
All available metric types must be known in order to generate a
generic search query, but these aren't available to the plugin until
the filter dropdown is activiated. To address this limitation we now
query for the available metric types when running a query if '*' is
selected; This also ensures a complete query if metrics were added
since the last request.

Because the web request method is async buildQueryParameters must be
async; The return type of this function is now a Promise.